Update babel-eslint to ^10.0.1
This "sounds bigger than it is": Basically we get a rebase of everything to babel v7, which
brings a conflicting change for "decorators" which we don't use for anything and therefore
can better start with the "official" ones.

babel-eslint 10.0.0 brought improved support for type aliases, and babel-eslint 10.0.1 reverted
that due to problems.

The update however bumps the minimum eslint version to 5, which dropped support for NodeJS 4.x.