Switch back to using eslint's default parser
eslint supports all "standard" features, babel-eslint would only help with "experimental" things. The last
experimental thing we used was the object rest spreading, which is part of ECMAScript 2018, and supported
in NodeJS since ~8.6.0[1].

Using babel-eslint has a bad side-effect though: It accepts TypeScript syntax such as specifying a type on a
variable, with seemingly no way to disable that.

This commit reverts cda318b, and bumps the parser ECMAScript version to 2018. The latter could also be reverted
and replaced with setting `parserOptions.ecmaFeatures.experimentalObjectRestSpread` to true, but that felt
less "clean".

[1] https://node.green/#ES2018-features-object-rest-spread-properties