Configure the new @typescript-eslint/naming-convention rule
The `camelcase` rule of eslint is still used for JavaScript sources, for TypeScript that rule
is deprecated (and gone) and we need to use the more fancy naming-convention rule. This does allow
to adjust this a bit further though, for example we can now require the `_`-convention for unused
variables.

While there: Extend the "tests" a bit to show off more of the conventions.